Ass-RT: add `size_t` and `bool`

+++ ass-std.h
@@ -32,7 +32,7 @@ extern int fclose(FILE* file);
 #define sprintf(b,f,rest...) TEXT->Format(b,f, ## rest)
 
 extern char * strtok_r (char *newstring, const char *delimiters, char **save_ptr);
-extern int strlen (char *s);
+extern size_t strlen (char *s);
 extern int strcmp (char *a, char *b);
 
 #endif
